+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Support approval macros for external auth header actions, including macro descriptors in pending webhooks and interpolation of approved values into per-rule header actions [#19]
 - Update external auth demo webapp to support approval macros end-to-end [#21]
+- Expose configurable external auth callbackUrl in webhooks and update demo webapp to consume it [#22]
 
 ## [0.0.2] - 2025-12-06
 
@@ -37,3 +38,4 @@
 [#2]: https://github.com/kcosr/acl-proxy/pull/2
 [#19]: https://github.com/kcosr/acl-proxy/pull/19
 [#21]: https://github.com/kcosr/acl-proxy/pull/21
+[#22]: https://github.com/kcosr/acl-proxy/pull/23